UI summary refactoring

+++ Energy.App.Standalone/Features/Setup/Store/ImmutatableStateObjects/HourOfDayPriceState.cs
@@ -4,6 +4,6 @@
     {
         public TimeSpan? HourOfDay { get; init; }
 
-        public double PencePerKWh { get; init; }
+        public decimal PencePerKWh { get; init; }
     }
 }
